Add registerLogHandler() to Python API
This is a first cut where all new code is confined to the SWIG
file.

registerLogHandler() can be called with a Python callback.
If any log messages that are generated prior or during an API call
then the callback is run for each message.  The messages are managed
in a threadsafe queue.